// src/types/types.ts
export type ApplicationCreateFormProps = {}
export type documentType = {
  id: number
  name: string
  format: string
}
export type productType = {
  id: number
  name: string
  able?: string
  docM: string
  docN: string
}
export type testProductType = {
  id: number
  name: string
  place?: string
  fio: string
  count: string
  nameL: string
  typeM: string
}